[@kadena/graph] Removed duplicate queries
