+++ .changeset/red-berries-own.md
@@ -1,0 +1,6 @@
+---
+'@kadena/graph-client': patch
+'@kadena/graph': patch
+---
+
+Removed duplicate queries
