+++ packages/apps/graph-client/src/pages/gas/estimation/index.tsx
@@ -9,6 +9,7 @@ import {
   BreadcrumbsItem,
   Cell,
   Column,
+  Notification,
   Row,
   Stack,
   Table,
@@ -51,40 +52,46 @@ const GasEstimation: React.FC = () => {
         loaderText="Waiting for gas limit estimation..."
       />
 
-      <Table isCompact className={atoms({ wordBreak: 'break-word' })}>
-        <TableHeader>
-          <Column>Type</Column>
-          <Column>Value</Column>
-        </TableHeader>
-        <TableBody>
-          <Row>
-            <Cell>Input</Cell>
-            <Cell>{router.query.input}</Cell>
-          </Row>
-          <Row>
-            <Cell>Gas Limit Estimate</Cell>
-            <Cell>{data?.gasLimitEstimate.amount}</Cell>
-          </Row>
-          <Row>
-            <Cell>The detected input type</Cell>
-            <Cell>{data?.gasLimitEstimate.inputType}</Cell>
-          </Row>
-          <Row>
-            <Cell>Was preflight used</Cell>
-            <Cell>{data?.gasLimitEstimate.usedPreflight.toString()}</Cell>
-          </Row>
-          <Row>
-            <Cell>Was signature verification used</Cell>
-            <Cell>
-              {data?.gasLimitEstimate.usedSignatureVerification.toString()}
-            </Cell>
-          </Row>
-          <Row>
-            <Cell>The transaction that was generated and used</Cell>
-            <Cell>{data?.gasLimitEstimate.transaction}</Cell>
-          </Row>
-        </TableBody>
-      </Table>
+      {data?.gasLimitEstimate?.length && data.gasLimitEstimate[0] ? (
+        <Table isCompact className={atoms({ wordBreak: 'break-word' })}>
+          <TableHeader>
+            <Column>Type</Column>
+            <Column>Value</Column>
+          </TableHeader>
+          <TableBody>
+            <Row>
+              <Cell>Input</Cell>
+              <Cell>{router.query.input}</Cell>
+            </Row>
+            <Row>
+              <Cell>Gas Limit Estimate</Cell>
+              <Cell>{data?.gasLimitEstimate[0].amount}</Cell>
+            </Row>
+            <Row>
+              <Cell>The detected input type</Cell>
+              <Cell>{data?.gasLimitEstimate[0].inputType}</Cell>
+            </Row>
+            <Row>
+              <Cell>Was preflight used</Cell>
+              <Cell>{data?.gasLimitEstimate[0].usedPreflight.toString()}</Cell>
+            </Row>
+            <Row>
+              <Cell>Was signature verification used</Cell>
+              <Cell>
+                {data?.gasLimitEstimate[0].usedSignatureVerification.toString()}
+              </Cell>
+            </Row>
+            <Row>
+              <Cell>The transaction that was generated and used</Cell>
+              <Cell>{data?.gasLimitEstimate[0].transaction}</Cell>
+            </Row>
+          </TableBody>
+        </Table>
+      ) : (
+        <Notification role="alert" intent="negative">
+          Unable to display gas limit estimation.
+        </Notification>
+      )}
     </>
   );
 };
