+++ packages/apps/graph/src/graph/index.ts
@@ -22,6 +22,7 @@ import './query/block';
 import './query/blocks-from-depth';
 import './query/blocks-from-height';
 import './query/completed-block-heights';
+import './query/event';
 import './query/events';
 import './query/fungible-account';
 // import './query/fungible-account-by-public-key';
@@ -33,8 +34,10 @@ import './query/last-block-height';
 import './query/non-fungible-account';
 import './query/non-fungible-chain-account';
 import './query/pact-query';
+import './query/transaction';
 import './query/transaction-by-public-key';
 import './query/transactions';
+import './query/transfer';
 import './query/transfers';
 import './subscription/events';
 import './subscription/new-blocks';
