+++ packages/apps/graph/src/graph/query/transactions.ts
@@ -4,34 +4,6 @@ import { getDefaultConnectionComplexity } from '@services/complexity';
 import { normalizeError } from '@utils/errors';
 import { builder } from '../builder';
 
-builder.queryField('transaction', (t) =>
-  t.prismaField({
-    description: 'Retrieve one transaction by its unique key.',
-    nullable: true,
-    args: {
-      blockHash: t.arg.string({ required: true }),
-      requestKey: t.arg.string({ required: true }),
-    },
-    type: Prisma.ModelName.Transaction,
-    complexity: getDefaultConnectionComplexity(),
-    async resolve(query, __parent, args) {
-      try {
-        return await prismaClient.transaction.findUnique({
-          ...query,
-          where: {
-            blockHash_requestKey: {
-              blockHash: args.blockHash,
-              requestKey: args.requestKey,
-            },
-          },
-        });
-      } catch (error) {
-        throw normalizeError(error);
-      }
-    },
-  }),
-);
-
 builder.queryField('transactions', (t) =>
   t.prismaConnection({
     description: 'Retrieve transactions. Default page size is 20.',
